I don't know, last shot

+++ include/aws/common/math.msvc.inl
@@ -84,7 +84,7 @@ AWS_STATIC_IMPL int aws_mul_u32_checked(uint32_t a, uint32_t b, uint32_t *r) {
  * a + b, returns the result in *r, and returns AWS_OP_SUCCESS.
  */
 AWS_STATIC_IMPL int aws_add_u64_checked(uint64_t a, uint64_t b, uint64_t *r) {
-    uint8_t c_in = (uint8_t)0;
+    uint8_t c_in = 0u;
     if (_addcarry_u64(c_in, a, b, r)) {
         return aws_raise_error(AWS_ERROR_OVERFLOW_DETECTED);
     }
@@ -96,7 +96,7 @@ AWS_STATIC_IMPL int aws_add_u64_checked(uint64_t a, uint64_t b, uint64_t *r) {
  */
 AWS_STATIC_IMPL uint64_t aws_add_u64_saturating(uint64_t a, uint64_t b) {
     uint64_t res = 0;
-    uint8_t c_in = (uint8_t)0;
+    uint8_t c_in = 0u;
 
     if (_addcarry_u64(c_in, a, b, &res)) {
         res = UINT64_MAX;
@@ -110,7 +110,7 @@ AWS_STATIC_IMPL uint64_t aws_add_u64_saturating(uint64_t a, uint64_t b) {
  * a + b, returns the result in *r, and returns AWS_OP_SUCCESS.
  */
 AWS_STATIC_IMPL int aws_add_u32_checked(uint32_t a, uint32_t b, uint32_t *r) {
-    uint8_t c_in = (uint8_t)0;
+    uint8_t c_in = 0u;
     if (_addcarry_u32(c_in, a, b, r)) {
         return aws_raise_error(AWS_ERROR_OVERFLOW_DETECTED);
     }
@@ -122,7 +122,7 @@ AWS_STATIC_IMPL int aws_add_u32_checked(uint32_t a, uint32_t b, uint32_t *r) {
  */
 AWS_STATIC_IMPL uint32_t aws_add_u32_saturating(uint32_t a, uint32_t b) {
     uint32_t res = 0;
-    uint8_t c_in = (uint8_t)0;
+    uint8_t c_in = 0u;
 
     if (_addcarry_u32(c_in, a, b, &res)) {
         res = UINT32_MAX;
